refactor(annotation): annotation
domain 폴더 내부에 있던 크린 아키텍처의 의미를 명확하게 해주는 어노테이션 모음을 common으로 이동하였습니다.
