+++ src/main/java/com/space/server/common/annotation/UseCase.java
@@ -1,7 +1,4 @@
-package com.space.server.domain.common;
-
-import org.springframework.core.annotation.AliasFor;
-import org.springframework.stereotype.Component;
+package com.space.server.common.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -9,13 +6,17 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.stereotype.Component;
+
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Component
-public @interface WebAdapter {
+public @interface UseCase {
 
   @AliasFor(annotation = Component.class)
   String value() default "";
 
 }
+
