+++ src/main/java/com/space/server/domain/inventory/application/service/GetInventoriesByIsEquippedService.java
@@ -1,6 +1,6 @@
 package com.space.server.domain.inventory.application.service;
 
-import com.space.server.domain.common.UseCase;
+import com.space.server.common.annotation.UseCase;
 import com.space.server.domain.inventory.application.port.in.GetInventoriesByIsEquippedQuery;
 import com.space.server.domain.inventory.application.port.out.LoadInventoriesByIsEquippedPort;
 import com.space.server.domain.inventory.domain.Inventory;
