+++ src/main/java/com/space/server/domain/inventory/application/service/GetInventoriesService.java
@@ -1,6 +1,6 @@
 package com.space.server.domain.inventory.application.service;
 
-import com.space.server.domain.common.UseCase;
+import com.space.server.common.annotation.UseCase;
 import com.space.server.domain.inventory.application.port.in.GetInventoriesQuery;
 import com.space.server.domain.inventory.application.port.out.LoadInventoriesPort;
 import com.space.server.domain.inventory.domain.Inventory;
