+++ src/main/java/com/space/server/domain/item/application/service/GetItemService.java
@@ -2,7 +2,7 @@ package com.space.server.domain.item.application.service;
 
 import com.space.server.domain.item.application.port.in.GetItemQuery;
 import com.space.server.domain.item.application.port.out.LoadItemPort;
-import com.space.server.domain.common.UseCase;
+import com.space.server.common.annotation.UseCase;
 import com.space.server.domain.item.domain.Item;
 import lombok.RequiredArgsConstructor;
